+ Moved all duplicate themes and moved them all to MainTheme.js
+ Changed theme. -> mainTheme. in the .jsx files that were affected by this migration

+ SignUp and SingIn are now the same theme.js file since they share everything.

- Moved SignUp.js and SignIn.js to deprecated folder.
