+++ frontend/src/components/AppBar.jsx
@@ -2,6 +2,7 @@ import { View, Pressable, Text } from 'react-native';
 import { Link } from 'react-router-native';
 import theme from '../themes/AppBarTheme';
 
+
 const AppBar = ({ user }) => {
     return (
         <View style={theme.container}>
