+++ frontend/src/components/WelcomeView.jsx
@@ -1,6 +1,8 @@
 import { Pressable, View, Text, ImageBackground } from "react-native";
 import { useNavigate } from "react-router-native";
 import theme from "../themes/WelcomeViewTheme";
+import mainTheme from "../themes/MainTheme";
+
 
 const WelcomeView = ({ user, onLogout }) => {
     const navigate = useNavigate();
@@ -15,9 +17,9 @@ const WelcomeView = ({ user, onLogout }) => {
     };
 
     return (
-        <View style={theme.blueContainer}>
-            <ImageBackground source={require('../../assets/BackgroundBinary.png')} style={theme.backgroundImage}>
-                <Text style={theme.titlePode}>{"<Pode/>"}</Text>
+        <View style={mainTheme.blueContainer}>
+            <ImageBackground source={require('../../assets/BackgroundBinary.png')} style={mainTheme.backgroundImage}>
+                <Text style={mainTheme.titlePode}>{"<Pode/>"}</Text>
                 <Pressable style={theme.button} onPress={onPressLogout}><Text>logout</Text></Pressable>
                 <Pressable style={theme.button} onPress={onPressLessons}><Text>lessons</Text></Pressable>
             </ImageBackground>
