+++ frontend/src/deprecated/SignUpTheme.js
