+++ frontend/src/themes/ExerciseToRenderTheme.js
@@ -9,13 +9,7 @@ const theme = {
         alignContent: "center",
         padding: 10,
       },
-      exerciseFont: {
-        color: "rgba(187,144,214,1)",
-        fontSize: 28,
-        fontWeight: "bold",
-        fontFamily: "AlfaSlabOne",
-        padding: 10,
-      },
+
 };
 
 export default theme;
