+++ frontend/src/themes/LessonViewTheme.js
@@ -1,11 +1,5 @@
 const LessonViewTheme = {
 
-    blueContainer: {  // Theme used in: LessonView
-        flexGrow: 1,
-        flexShrink: 1,
-        backgroundColor: "rgba(127,222,255,1)",
-        justifyContent: "center",
-      },
       whiteContainer: {     // Theme used in: LessonView
         marginVertical: 0,
         marginHorizontal: 18,
@@ -53,14 +47,6 @@ const LessonViewTheme = {
       letsCodeButtonContainer: {    // Theme used in: LessonView
         flexDirection: "column",
         justifyContent: "center",
-      },
-      greenButton: {    // Theme used in: LessonView
-        margin: 20,
-        padding: 10,
-        backgroundColor: "#84DC95",
-        borderRadius: 33,
-        borderWidth: 6,
-        borderColor: "#4A8055",
       },
       letsCodeText: {   // Theme used in: LessonView
         fontFamily: "AlfaSlabOne",
