+++ frontend/src/themes/SignInAndSignUpTheme.js
@@ -1,0 +1,58 @@
+const theme = {
+
+      pinkContainer: {
+        marginHorizontal: 18,
+        backgroundColor: "rgba(237,220,249,1)",
+        borderColor: "rgba(187,144,214,1)",
+        borderRadius: 48,
+        borderWidth: 7,
+        alignContent: "center",
+        padding: 10,
+      },
+      arrowContainer: {
+        height: 30,
+        width: 30,
+        margin: 10,
+      },
+      arrow: {
+        fontSize: 30,
+        color: "rgba(75,113,123,1)",
+      },
+      titlePode: {
+        color: "rgba(75,113,123,1)",
+        fontSize: 70,
+        textAlign: "center",
+        paddingVertical: 10,
+        fontFamily: "AlfaSlabOne",
+      },
+      errorText: {
+        color: "#d73a4a",
+        marginLeft: 10
+      },
+      inputField: {
+        paddingVertical: 10,
+        backgroundColor: "rgba(255,255,255,1)",
+        borderColor: "rgba(100,152,166,1)",
+        borderRadius: 10,
+        borderWidth: 2,
+        color: "rgba(100,152,166,1)",
+        fontSize: 25,
+      },
+      purpleButton: {
+        margin: 20,
+        padding: 10,
+        backgroundColor: "rgba(187,144,214,1)",
+        borderRadius: 33,
+        borderWidth: 6,
+        borderColor: "rgba(135,105,155,1)",
+        color: "BB90D6",
+      },
+      purpleButtonText: {
+        color: "rgba(135,105,155,1)",
+        fontSize: 45,
+        textAlign: "center",
+        fontFamily: "AlfaSlabOne",
+      },
+};
+
+export default theme;
