+++ frontend/src/themes/SingleExerciseViewTheme.js
@@ -1,10 +1,5 @@
 const theme = {
-    blueContainer: { 
-        flexGrow: 1,
-        flexShrink: 1,
-        backgroundColor: "rgba(127,222,255,1)",
-        justifyContent: "center",
-      },
+
       whiteContainerExercises: {
         marginVertical: 18, 
         marginHorizontal: 18, 
@@ -17,43 +12,7 @@ const theme = {
         flex: 1,
       },
 
-      exerciseDescription: { 
-        color: "rgba(75,113,123,1)",
-        fontSize: 18,
-        textAlign: "left",
-        paddingVertical: 10,
-        paddingHorizontal: 10,
-        fontFamily: "AlfaSlabOne",
-        lineHeight: 24,
-      },
-      greenButton: {   
-        margin: 20,
-        padding: 10,
-        backgroundColor: "#84DC95",
-        borderRadius: 33,
-        borderWidth: 6,
-        borderColor: "#4A8055",
-      },
-      greenButtonText: {
-        color: "#4A8055",
-        fontSize: 45,
-        textAlign: "center",
-        fontFamily: "AlfaSlabOne",
-      },
-      greenButtonDeselected: {
-        margin: 20,
-        padding: 10,
-        backgroundColor: "#C2D6C6",
-        borderRadius: 33,
-        borderWidth: 6,
-        borderColor: "#A2B1A5",
-      },    
-      greenButtonTextDeselected: {
-        color: "#A2B1A5",
-        fontSize: 45,
-        textAlign: "center",
-        fontFamily: "AlfaSlabOne",
-      },
+
       container: {
         padding: 20,
     },
