+++ frontend/src/themes/WelcomeViewTheme.js
@@ -1,22 +1,6 @@
 const theme = {
-    blueContainer: {  // Theme used in: LessonView
-        flexGrow: 1,
-        flexShrink: 1,
-        backgroundColor: "rgba(127,222,255,1)",
-        justifyContent: "center",
-      },
-      backgroundImage: {
-        flex: 1,
-        resizeMode: 'cover',
-        justifyContent: 'center',
-      },
-      titlePode: {
-        color: "rgba(75,113,123,1)",
-        fontSize: 70,
-        textAlign: "center",
-        paddingVertical: 10,
-        fontFamily: "AlfaSlabOne",
-      },
+
+
       button: {
         backgroundColor: '#0366d6',
         alignItems: "center",
